Individual and population included. The initial population is of size 40 and their cromosome is generated randomly. Their fitness value is also calculated.

+++ src/Game.hpp
@@ -38,7 +38,7 @@ class logic::Game
 	bool isRoundOver;
 
 	// Stores the score of the game.
-	float score;
+	double score;
 
 	/* Stores the map coordinates of the shortest path from the respawn point 
 	 	to the player tower */
